+++ CHANGES.md
@@ -1,5 +1,6 @@
 **Added:**
 
+- A Git repo with an initial commit is now created by default when creating a new project. Pass a `--no-git` flag to disable this.
 - Added an `audio` loader.
 
 **Removed:**